+++ src/main/php/net/stubbles/lang/types/datespan/AbstractDatespan.php
@@ -60,6 +60,30 @@ abstract class AbstractDatespan implements Datespan
     }
 
     /**
+     * checks whether datespan starts before a given date
+     *
+     * @param   int|string|\DateTime|Date  $date
+     * @return  bool
+     * @since   3.5.0
+     */
+    public function startsBefore($date)
+    {
+        return $this->start->isBefore($date);
+    }
+
+    /**
+     * checks whether datespan starts after a given date
+     *
+     * @param   int|string|\DateTime|Date  $date
+     * @return  bool
+     * @since   3.5.0
+     */
+    public function startsAfter($date)
+    {
+        return $this->start->isAfter($date);
+    }
+
+    /**
      * returns the end date
      *
      * @return  Date
@@ -67,6 +91,56 @@ abstract class AbstractDatespan implements Datespan
     public function getEnd()
     {
         return $this->end;
+    }
+
+    /**
+     * checks whether datespan ends before a given date
+     *
+     * @param   int|string|\DateTime|Date  $date
+     * @return  bool
+     * @since   3.5.0
+     */
+    public function endsBefore($date)
+    {
+        return $this->end->isBefore($date);
+    }
+
+    /**
+     * checks whether datespan ends after a given date
+     *
+     * @param   int|string|\DateTime|Date  $date
+     * @return  bool
+     * @since   3.5.0
+     */
+    public function endsAfter($date)
+    {
+        return $this->end->isAfter($date);
+    }
+
+    /**
+     * returns formatted date/time string for start date
+     *
+     * @param   string    $format    format, see http://php.net/date
+     * @param   TimeZone  $timeZone  target time zone of formatted string
+     * @return  string
+     * @since   3.5.0
+     */
+    public function formatStart($format, TimeZone $timeZone = null)
+    {
+        return $this->start->format($format, $timeZone);
+    }
+
+    /**
+     * returns formatted date/time string for end date
+     *
+     * @param   string    $format    format, see http://php.net/date
+     * @param   TimeZone  $timeZone  target time zone of formatted string
+     * @return  string
+     * @since   3.5.0
+     */
+    public function formatEnd($format, TimeZone $timeZone = null)
+    {
+        return $this->end->format($format, $timeZone);
     }
 
     /**
@@ -123,4 +197,3 @@ abstract class AbstractDatespan implements Datespan
         return $this->asString();
     }
 }
-?>